Separate out LoadConfiguration callable
Shrink the class so we can add more methods.
