+++ lib/baes.rb
@@ -5,6 +5,7 @@ module Baes; end
 
 class Baes::Error < StandardError; end
 
+require_relative "baes/actions"
 require_relative "baes/configuration"
 require_relative "baes/branch"
 require_relative "baes/git"
