+++ lib/baes/actions.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# namespace for callable modules
+module Baes::Actions; end
+
+require_relative "actions/load_configuration"
